requirements.txt
async-generator==1.10
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
google-auth==2.6.0
google-auth-oauthlib==0.5.0
gspread==5.2.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
oauthlib==3.2.0
outcome==1.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
webdriver-manager==3.5.3
wsproto==1.1.0
